Bring game cookie and DB structure closer together
Add a topics_string to each category in the database. While this
slightly denormalizes the database, the previous solution of each
category having first_topic and last_topic fields was absurdly hackish,
not to mention making more work of everything relating to topics, for
little benefit.

Add id and board_position fields to the game cookie. This should make
things much easier when saving games.