Correct date field values for existing games
Grab values from the database or cookie, if appropriate, and set the
show-date, played-at, and play-type fields appropriately.

If there is no database request and no existing cookie (or if the
delete-game link is clicked), populate these values with a best guess
(most-recent show's date and play type, current time).