Update topics on save
When a user saves a game, go through each category's topics_string,
convert it to an array of topic names, and convert that to an array of
topic objects (using first_or_create! to grab a preexisting topic or
create a new one, as appropriate).

Set the category's topics equal to the aforementioned array of topic
objects.