Fix display bug on game edit in production
In production, if a user tries to edit a game via its link on the stats
page, the game will be loaded into the cookie properly, but most of the
time the game board will still be displaying whatever was there
previously. This bug was not present in the development environment.

The issue is that the populateBoard() function was being called
immediately, rather than queuing it in the success function of the Ajax
request. In development, with the server on the same machine, this
request was fulfilled so quickly that no problems arose.

However, in production, with the server elsewhere, the lag was long
enough that populateBoard() was being called before the data that it
needed to be populated with was available. This commit moves that call
into the Ajax request's success function, so the board will not be
populated until the proper data is present.