Prevent errors from unorthodox topics_strings
Previously, if the user had two commas with no non-whitespace characters
in between, Topicable would attempt to create a new topic with the empty
string as a name, which raised an error and caused the entire save to
fail. Remedy this by only adding a topic to the list if it's valid.

Also, if the user entered the same topic name twice in the same string,
this would cause the save to fail, due to violation of the topic-name
uniqueness constraint. Remove any duplicates (case-insensitively) from
the array before hitting the database.