Make date fields responsive
Previously, the "show date" and "played at" fields were the only parts
of the game page that didn't respond to the size of the viewport. This
led to them being fairly small in big windows and format-breakingly
large in small windows.

This commit makes these fields adjust to the viewport size.

It also aligns the dates and their headings, and (almost) aligns the
play type list with its heading, for increased visual pleasure.