Add error alert on Ajax request during maintenance
Previously, when a save or date change was requested with the app in
maintenance mode, the response (a Heroku error page with status 503)
was not anticipated by the error handler functions, thereby leaving the
link hanging in the "Saving..." or "changing..." state.

This commit adds a check for 503 status and displays an appropriate
alert when the app is in maintenance mode.