Enable undoing of game reset
Previously, when the "New game" link was clicked, the current game
vanished (irrevocably if unsaved).

Now, when the "New game" link is clicked, it turns into an "Undo" link,
which will restore the previous game. (The "Undo" link turns back into
a "New game" link as soon as any change is made to the new game.)